+++ tests/screenshot/apps/importer.js
@@ -8,6 +8,7 @@ import CheckboxItem from '../../../CheckboxItem';
 import ContextualMenuDecorator from '../../../ContextualMenuDecorator';
 import ContextualPopupDecorator from '../../../ContextualPopupDecorator';
 import DatePicker from '../../../DatePicker';
+import DayPicker from '../../../DayPicker';
 import Dropdown from '../../../Dropdown';
 import FixedPopupPanels from '../../../FixedPopupPanels';
 import FlexiblePopupPanels from '../../../FlexiblePopupPanels';
@@ -54,6 +55,7 @@ const components = {
 	ContextualMenuDecorator,
 	ContextualPopupDecorator,
 	DatePicker,
+	DayPicker,
 	Dropdown,
 	FixedPopupPanels,
 	FlexiblePopupPanels,